+++ atouin/src/main/kotlin/dlm/entities/elements/ColorMultiplicator.kt
@@ -1,12 +1,12 @@
 package dlm.entities.elements
 
 data class ColorMultiplicator(
-    val red: Byte,
-    val green: Byte,
-    val blue: Byte,
+    val red: Double,
+    val green: Double,
+    val blue: Double,
     val isAlreadyComputed: Boolean = false
 ) {
     companion object {
-        fun clamp(value: Int, min: Double, max: Double): Double = value.toDouble().coerceIn(min, max)
+        fun clamp(value: Double, min: Double, max: Double): Double = value.coerceIn(min, max)
     }
 }
