+++ atouin/src/main/kotlin/dlm/services/elements/GraphicalElementService.kt
@@ -16,8 +16,16 @@ class GraphicalElementService : ParamsParserService<GraphicalElement, Cell> {
         val result = GraphicalElement(params)
 
         result.elementId = raw.readUnsignedInt()
-        result.hue = ColorMultiplicator(raw.readByte(), raw.readByte(), raw.readByte())
-        result.shadow = ColorMultiplicator(raw.readByte(), raw.readByte(), raw.readByte())
+        result.hue = ColorMultiplicator(
+            raw.readByte().toDouble(),
+            raw.readByte().toDouble(),
+            raw.readByte().toDouble()
+        )
+        result.shadow = ColorMultiplicator(
+            raw.readByte().toDouble(),
+            raw.readByte().toDouble(),
+            raw.readByte().toDouble()
+        )
         if (params.layer.map.mapVersion <= 4) {
             result.offset = Point2D.Double(
                 raw.readByte().toDouble(),
@@ -48,7 +56,7 @@ class GraphicalElementService : ParamsParserService<GraphicalElement, Cell> {
         val g = ColorMultiplicator.clamp((params.hue.green + params.shadow.green + 128) * 2, 0.0, 512.0)
         val b = ColorMultiplicator.clamp((params.hue.blue + params.shadow.blue + 128) * 2, 0.0, 512.0)
         params.finalTeint = ColorMultiplicator(
-            r.toInt().toByte(), g.toInt().toByte(), b.toInt().toByte(),
+            r, g, b,
             isAlreadyComputed = true
         )
     }
